+++ src/features/MainScreen/index.tsx
@@ -6,9 +6,9 @@ import { useTagsRegistry } from '@features/App/Workspace/WorkspaceProvider';
 import { NotesListPanel } from '@features/MainScreen/NotesListPanel';
 import { WorkspacesPanel } from '@features/MainScreen/WorkspacesPanel';
 import { NotesContainer } from '@features/NotesContainer';
-import { useWorkspaceShortcutHandlers } from '@hooks/commands/shortcuts/useWorkspaceShortcutHandlers';
 import { useNoteCommandHandlers } from '@hooks/notes/useNoteCommandHandlers';
 import { useUpdateNotes } from '@hooks/notes/useUpdateNotes';
+import { useWorkspaceShortcutHandlers } from '@hooks/useWorkspaceShortcutHandlers';
 
 import { NotificationsPopup } from '../NotificationsPopup';
 import { ActivityBar } from './ActivityBar';
